Allows the user to enable/disable the dependency between the godotRunEditor task and the build task.
The build task may take a fair amount of time; not ideal for quickly starting the editor. If
the build currently is broken (e.g. compile errors) it may prevent the user from starting the
editor at all, which leads to a hunt for the executable. Let's avoid this situation.